\name{AddLine}
\alias{AddLine}
\title{AddLine}
\description{Add a line to the current plotting area.}
\arguments{
\item{intercept}{Intercept of the line in the current plot coordinate system}
\item{slope}{Slope of the line in the current plot coordinate system}
\item{originLeft}{X coordinate to start drawing the line from (default: NULL)}
\item{originRight}{Y coordinate to end drawing the line at (default: NULL)}
\item{trimLeft}{Given a linelist, which point should be the line start from (default: NULL)}
\item{trimRight}{Given a linelist, which point should be the line end at (default: NULL)}
\item{linelist}{Set of points formatted as required for LinearPlot (default: NULL)}
\item{...}{}
}
\author{Matt Arnold}
