\name{AxisLabels}
\alias{AxisLabels}
\title{AxisLabels}
\description{Add X and Y axis labels to the current panel.}
\arguments{
\item{xlab}{X axis label}
\item{ylab}{Y axis label}
\item{mainfont}{Font scaling applied to cex (default: NULL)}
\item{xline}{Margin line of the X axis label (default: 3.5)}
\item{yline}{Margin line of the Y axis label (default: 4)}
\item{cex}{Font scaling factor (multiplies the underlying panel "unit" font size) (default: NULL)}
\item{adj}{Alignment of the axis labels, as adj in text() (default: 0.5)}
}
\author{Matt Arnold}
