\name{blankPlot}
\alias{blankPlot}
\title{blankPlot}
\description{Create an empty plot in the next available panel, which also sets up a user-coordinate system. This allows further arbitrary drawing commands to be sent to the panel.}
\arguments{
\item{xlim}{Vector giving min and max value of X range to be plotted}
\item{ylim}{Vector giving min and max value of Y range to be plotted}
\item{mainfont}{Font scaling applied to cex (default: 1)}
\item{XLogScale}{ (default: FALSE)}
\item{YLogScale}{ (default: FALSE)}
}
\author{Matt Arnold}
