\name{closeFile}
\alias{closeFile}
\title{closeFile}
\description{Close the currently active file, if one is open.}
\arguments{
\item{type}{If type is "GUI" then ignore the request to close the file, all other values close the currently open file (default: GUI)}
\item{suppress.notice}{Turn off the notice that a file has been closed (default: FALSE)}
}
\author{Matt Arnold}
