\name{convertLineToUser}
\alias{convertLineToUser}
\title{convertLineToUser}
\description{Convert margin line number to a user coordinate. Useful for aligning plot labels.}
\arguments{
\item{line}{Margin line to be converted into the user coordinate}
\item{side}{Margin side, one of 1, 2, 3, 4 corresponding to  bottom, left, top, right}
}
\author{Matt Arnold}
