\name{FindCSVs}
\alias{FindCSVs}
\title{FindCSVs}
\description{Find and load all the CSV files in a particular folder. The CSVs are loaded into data frames which have the same name as the corresponding filestem.}
\arguments{
\item{folder}{Path to folder (default: .)}
\item{verbose}{ (default: TRUE)}
\item{upcase.names}{ (default: FALSE)}
}
\author{Matt Arnold}
