\name{mainTitle}
\alias{mainTitle}
\title{mainTitle}
\description{Add a title to the current plot, in the title region. This function attempts to make the title as large as possible within the constraints of the title space allocated.}
\arguments{
\item{titles}{Vector of titles to put in the title space allocated on the page. First element is larger than the remainder}
\item{mainfont}{Font scaling applied to cex (default: NULL)}
\item{shrinktofit}{Shrink the font size automatically until the widest title line fits in the space allocated (default: TRUE)}
\item{cex}{Font scaling factor (multiplies the underlying panel "unit" font size) (default: NULL)}
}
\author{Matt Arnold}
