\name{saveGUI}
\alias{saveGUI}
\title{saveGUI}
\description{Saves the contents of the currently active GUI window to a file.}
\arguments{
\item{filename}{Filename stem, the correct extension (depending on the value of type) will be automatically appended}
\item{type}{Only PDF currently supported (default: pdf)}
\item{x}{Width of the plot in the file (default: 6)}
\item{y}{Height of the plot in the file (default: 6)}
}
\author{Matt Arnold}
