\name{WLScoefs}
\alias{WLScoefs}
\title{WLScoefs}
\description{Computed weighted least-squares coefficients for a LinearPlot-formatted list. The result is a matrix which can be used as the reglines argument in LinearPlot.}
\arguments{
\item{linelist}{List of lines formatted as required for LinearPlot}
\item{DataLogged}{Data is stored in log format in the data frame (default: FALSE)}
}
\author{Matt Arnold}
