\name{writePvalue}
\alias{writePvalue}
\title{writePvalue}
\description{Write out a number in `a times 10 to the power of b' notation.}
\arguments{
\item{pvalues}{Numeric vector of values to be written out using superscript for power of 10}
\item{x}{X-axis coordinates for labels}
\item{y}{Y-axis coordinates for labels}
\item{use.lower.limit}{ (default: FALSE)}
\item{lower.limit}{ (default: 1e-100)}
\item{use.upper.limit}{ (default: FALSE)}
\item{upper.limit}{ (default: 0.01)}
\item{cex}{Font scaling factor (multiplies the underlying panel "unit" font size) (default: 1)}
\item{col}{Font colour (default: 1)}
\item{adj}{Position adjustment (as in the text() command) (default: 0.5)}
\item{font}{ (default: 1)}
}
\author{Matt Arnold}
