\name{Xaxis}
\alias{Xaxis}
\title{Xaxis}
\description{Add an X axis to the current panel.}
\arguments{
\item{xticks}{Vector of tick marks to put on the X axis}
\item{ticklabs}{Character vector of tick labels to use instead of computed numeric values (default: NULL)}
\item{mainfont}{Font scaling applied to cex (default: 1)}
\item{VerticalTickLabs}{Write the tick labels perpendicular to the axis (default: FALSE)}
\item{BootLeft}{deprecated (default: FALSE)}
\item{BootRight}{deprecated (default: FALSE)}
\item{lwd}{Line width (scaling) (default: 1)}
\item{allX}{Extend the axis to cover the whole range of the current panel (default: TRUE)}
\item{line}{ (default: 1)}
\item{labels.line}{ (default: 1)}
\item{tcl}{ (default: -0.5)}
}
\author{Matt Arnold}
