\name{Yaxis}
\alias{Yaxis}
\title{Yaxis}
\description{Add a Y axis to the current panel.}
\arguments{
\item{yticks}{Vector of tick marks to put on the Y axis}
\item{LinScale}{ (default: TRUE)}
\item{ticklabs}{Character vector of tick labels to use instead of computed numeric values (default: NULL)}
\item{mainfont}{Font scaling applied to cex (default: 1)}
\item{side}{ (default: 2)}
\item{lwd}{Line width (scaling) (default: 1)}
\item{col}{ (default: 1)}
\item{line}{ (default: 1)}
\item{allY}{ (default: TRUE)}
}
\author{Matt Arnold}
